add hived switch
