+++ src/rest-server/src/config/v2/protocol.js
@@ -20,7 +20,7 @@
 const Ajv = require('ajv');
 const ajvMerge = require('ajv-merge-patch/keywords/merge');
 
-const ajv = new Ajv({allErrors: true, useDefaults: true});
+const ajv = new Ajv({allErrors: true});
 ajvMerge(ajv);
 
 // base schema
@@ -43,40 +43,6 @@ const baseSchema = {
     },
     description: {
       type: 'string',
-    },
-  },
-};
-
-// hived schema
-const hivedSchema = {
-  type: 'object',
-  properties: {
-    jobPriorityClass: {
-      enum: ['prod', 'test'],
-    },
-    taskRoles: {
-      patternProperties: {
-        '^[A-Za-z0-9._~]+$': {
-          type: 'object',
-          properties: {
-            gpuType: {
-              type: ['string', 'null'],
-              default: null,
-            },
-            reservationId: {
-              type: ['string', 'null'],
-              default: null,
-            },
-            affinityGroupName: {
-              type: ['string', 'null'],
-              default: null,
-            },
-          },
-          additionalProperties: false,
-        },
-      },
-      minProperties: 1,
-      additionalProperties: false,
     },
   },
 };
@@ -357,10 +323,6 @@ const protocolSchema = {
     },
     extras: {
       type: 'object',
-      properties: {
-        hivedScheduler: hivedSchema,
-      },
-      additionalProperties: true,
     },
   },
   required: [
